# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: E-Commerce Backend API
  version: 1.0.0
  description: |
    REST API for an e-commerce store: authentication, catalog, cart, orders,
    Stripe payments, and purchase-verified reviews.

    ## Response envelope
    Every JSON endpoint (except the Stripe webhook) responds with:

    | Outcome | Shape |
    |---|---|
    | Success | `{ "success": true, "message": string, "data": <payload or null> }` |
    | Error   | `{ "success": false, "message": string, "errorCode": string }` |

    `errorCode` is stable and machine-readable — branch on it, not on `message`.

    ## Authentication
    Send the access token from `/auth/login` or `/auth/register` as
    `Authorization: Bearer <accessToken>`. Access tokens expire after 15 minutes;
    exchange the refresh token (valid 7 days) at `/auth/refresh` for a new one.

    Each operation states its requirement: **Public**, **Authenticated** (any
    logged-in user), or **Admin only**. Admin-only operations are also marked
    with `x-required-role: ADMIN`.

    ## Errors common to every endpoint
    | Status | errorCode | When |
    |---|---|---|
    | 400 | `VALIDATION_ERROR` | Body, query, or path params failed validation. `message` lists each failing field as `location.field: reason`. |
    | 401 | `UNAUTHORIZED` / `INVALID_ACCESS_TOKEN` | Missing, malformed, or expired access token (protected endpoints only). |
    | 403 | `FORBIDDEN` | Authenticated but not allowed (wrong role, or not the owner). |
...

...

...
servers:
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Get a new access token
      description: |
        **Public** (the refresh token is the credential). Exchanges a valid,
        unrevoked refresh token for a new access token. The refresh token itself
        is not rotated and stays valid until it expires or is logged out.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenBody'
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: object
                        required: [accessToken]
                        additionalProperties: false
                        properties:
                          accessToken:
                            type: string
                            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIwYjZmN2MxZSJ9.Yc9mB2
              example:
                success: true
                message: Token refreshed
                data:
                  accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIwYjZmN2MxZSJ9.Yc9mB2
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Refresh token is malformed, expired, revoked (logged out), or its user no longer exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                revoked:
                  summary: Logged out / expired
                  value:
                    success: false
                    message: Refresh token is no longer valid
                    errorCode: INVALID_REFRESH_TOKEN
                malformed:
                  summary: Not a valid refresh JWT
                  value:
                    success: false
                    message: Invalid or expired refresh token
                    errorCode: INVALID_REFRESH_TOKEN
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/auth/login` or `/auth/register` (valid 15 minutes).
...
  responses:
...
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            message: 'body.email: Invalid email address, body.password: Password must be at least 8 characters'
            errorCode: VALIDATION_ERROR
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success: { const: true }
        message: { type: string, example: Success }
        data: {}
    ErrorResponse:
      type: object
      required: [success, message, errorCode]
      additionalProperties: false
      properties:
        success: { const: false }
        message: { type: string, example: Product not found }
        errorCode: { type: string, example: PRODUCT_NOT_FOUND }
...
    RefreshTokenBody:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIwYjZmN2MxZSIsImp0aSI6IjdhMWMifQ.pQ3r8Zt0sL
...